requirements.txt
streamlit
chromadb
openai
python-dotenv